move extension recommendtation out of platform
